Change the spinlock primitives to function as compiler barriers.
Backported from 0709b7ee72e4bc71ad07b7120acd117265ab51d0 with minor
conflict that arose due to the absence of sparcv8 as a platform.

Original commit message follows:

Previously, they functioned as barriers against CPU reordering but not
compiler reordering, an odd API that required extensive use of volatile
everywhere that spinlocks are used.  That's error-prone and has negative
implications for performance, so change it.

In theory, this makes it safe to remove many of the uses of volatile
that we currently have in our code base, but we may find that there are
some bugs in this effort when we do.  In the long run, though, this
should make for much more maintainable code.

Patch by me.  Review by Andres Freund.