Add a memory barrier in Own/DisownLatch
Failing to do so can lead to a race condition between an exiting backend
(Y) and another backend (X) undergoing init. If the act of returning the PGPROC
to the freelist for Y gets reordered before the store for procLatch->pid
= 0, then there is a chance that X can grab the same PGPROC object from
the freelist and can still see that procLatch->pid = Y. This can lead to
the PANIC inside OwnLatch(). This can only happen if X's OwnLatch() is
scheduled before Y's store for pid = 0.

Using a memory barrier in DisownLatch() is sufficient to avoid this
issue. Technically, the SpinLockAcquire() and Release() that succeeds
the DisownLatch() call should act as compiler barriers and should
suffice on x86, we still take this precautionary measure to be doubly
sure. The memory barrier inside OwnLatch() is an extension of that
paranoia. Since OwnLatch() and DisownLatch() aren't in hot code paths,
the added memory barriers should not be a performance sink.

Co-authored-by: Brent Doil <bdoil@vmware.com>
Co-authored-by: Jimmy Yih <jyih@vmware.com>
Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>